Crate oak_restricted_kernel_interface

Re-exports

pub use errno::Errno;

pub use syscalls::Syscall;

Modules

errno

initial_data

syscall

syscalls

Macros

syscall

Invoke system calls based on the Linux calling convention using SYSCALL.

Constants

DERIVED_KEY_FD

Predefined file descriptor for reading a derived key.

DICE_DATA_FD

Predefined file descriptor for reading the dice attestation data.

EVENT_LOG_FD

Predefined file descriptor for reading the eventlog attestation data.

OAK_CHANNEL_FD

Predefined file descriptor for the Oak communication channel.