Macro oak_restricted_kernel_interface::syscall

macro_rules! syscall {
    ($syscall:expr) => { ... };
    ($syscall:expr, $arg1:expr) => { ... };
    ($syscall:expr, $arg1:expr, $arg2:expr) => { ... };
    ($syscall:expr, $arg1:expr, $arg2:expr, $arg3:expr) => { ... };
    ($syscall:expr, $arg1:expr, $arg2:expr, $arg3:expr, $arg4:expr, $arg5:expr, $arg6:expr) => { ... };
}

Invoke system calls based on the Linux calling convention using SYSCALL.

See https://github.com/torvalds/linux/blob/master/arch/x86/entry/calling.h for the definition of the calling convention, and the AMD64 Architecture Programmer’s Manual, Volume 3 for the semantics of SYSCALL/SYSRET.