Enum oak_restricted_kernel_interface::syscalls::Syscall

#[repr(usize)]
pub enum Syscall {
    Read,
    Write,
    Mmap,
    Exit,
    Fsync,
}

System calls implemented by Oak Restricted Kernel.

In general, the system calls are inpired by and look similar to the Linux/POSIX ABI, but we make no claims about adhering to the behaviour, or specification, of either of those.

See https://github.com/torvalds/linux/blob/master/arch/x86/entry/syscalls/syscall_64.tbl for a full list of system call numbers; we will ever only support a small subset of these.

Variants

Read

Read from a file descriptor.

Arguments:

arg0 (c_ssize_t): file descriptor number

arg1 (*mut c_void): pointer to the buffer to be filled

arg2 (c_size_t): size of the buffer Returns: a value of errno::Errno on failure; otherwise, the number of bytes read.

Write

Write to a file descriptor.

Arguments:

arg0 (c_ssize_t): file descriptor number

arg1 (*const c_void): pointer to the buffer containing data to be written

arg2 (c_size_t): size of the buffer Returns: a value of errno::Errno on failure; otherwise, the number of bytes written.

Mmap

Creates a mapping for memory. Arguments:

arg0 (*const c_void): hint for start address for the new mapping, may be nullptr

arg1 (c_size_t): size of the new mapping

arg2 (c_int): protection on mapping (PROT_EXEC, PROT_READ, PROT_WRITE, PROT_NONE)

arg3 (c_int): flags. We require MAP_PRIVATE and MAP_ANONYMOUS to be set, and additionally support MAP_FIXED.

arg4 (c_int): file descriptor. Ignored, as we only support anonymous mappings. Should be set to -1 by caller.

arg5 (c_int): offset. Ignored, as we only support anonymous mappings. Should be set to 0 by caller. Oak Restricted Kernel considerations:

our mmap will work on 2 MiB chunks; even if size is 1 byte, we will reserve a 2 MiB chunk of memory. Thus, size should be kept as a multiple of 2 MiB.

related to previous, the allocation address will always be 2 MiB-aligned (rounded upward from hint).

MAP_FIXED requires address to be 2 MiB-aligned, and will return an error if it’d touch any existing mappings.

We do not support PROT_NONE; PROT_READ is always implied.

Exit

Terminates he calling process. Arguments:

arg0 (c_int): error code Oak Restricted Kernel considerations: We don’t expect the user process to terminate, so this triggers a kernel panic, no matter the error code.

Fsync

Flush a file descriptor. Arguments:

arg0 (c_ssize_t): file descriptor number. Ignored by the kernel as we don’t support multiple file descriptors. Returns: a value of errno::Errno on failure; 0, otherwise.

Implementations

impl Syscall

pub const fn from_repr(discriminant: usize) -> Option<Syscall>

Try to create Self from the raw representation